requirements.txt
django=4.0.6
djangorestframework
requests